USB: core: Enable root_hub's remote wakeup for wakeup sources
Now we only enable the remote wakeup function for the USB wakeup source
itself at usb_port_suspend(). However on some Loongson hosts this isn't
enough to enable the S3 wakeup function for USB keyboards, so we also
enable the root_hub's remote wakeup (and disable it on error). Frankly
this is a platform-specific problem, but enable it unconditionally make
code simple and seems harmless.

Change-Id: Ib86cc5cd331891b464d522ea00792a75dea1d4ce
Cc: stable@vger.kernel.org
Signed-off-by: Huacai Chen <chenhuacai@loongson.cn>